+++ pjproject/pjsip-apps/src/pjsua2/ios-swift-pjsua2/ios-swift-pjsua2/PJSua2/CustomPJSUA2.hpp
@@ -95,12 +95,12 @@ public:
     /**
      Hold the call
      */
-    void holdCall(int passid);
+    void holdCall(long passid);
     
     /**
      unhold the call
      */
-    void unholdCall(int passid);
+    void unholdCall(long passid);
     
     
     /**
